requirements/base.txt
Django==1.9.5
django-model-utils==2.5
djangorestframework==3.3.3
psycopg2==2.6.1
python-dateutil==2.5.3
dj-static==0.0.6
redis==2.10.5
django-redis-cache==1.6.5
django-tastypie==0.13.3

